+++ maiacore/src/maiacore/note.cpp
@@ -1558,6 +1558,36 @@ int Note::getScaleDegree(const Key& key) const {
 
 float Note::getFrequency() const { return Helper::pitch2freq(getSoundingPitch()); }
 
+std::pair<std::vector<float>, std::vector<float>> Note::getHarmonicSpectrum(
+    const int numPartials,
+    const std::function<std::vector<float>(std::vector<float>)> amplCallback) const {
+    if (numPartials <= 0) {
+        LOG_ERROR("The 'numPartials' must be a positive value");
+    }
+
+    std::vector<float> freqs(numPartials, 0);
+    for (int i = 0; i < numPartials; i++) {
+        freqs[i] = getFrequency() * (i + 1);
+    }
+
+    std::vector<float> ampls(numPartials, 0);
+    if (amplCallback == nullptr) {
+        for (int i = 0; i < numPartials; i++) {
+            ampls[i] = std::powf(0.88f, i);
+        }
+    } else {
+        ampls = amplCallback(freqs);
+    }
+
+    if (freqs.size() != ampls.size()) {
+        LOG_ERROR(
+            "The output vector of 'amplCallback' function must have the size of 'numPartials'=" +
+            std::to_string(numPartials));
+    }
+
+    return {freqs, ampls};
+}
+
 void Note::transpose(const int semitones, const std::string& accType) {
     const std::string newPitch = Helper::transposePitch(getPitch(), semitones, accType);
     setPitch(newPitch);
@@ -1839,7 +1869,7 @@ const std::string Note::getWrittenPitch() const {
 
 int Note::getWrittenOctave() const { return _writtenOctave; }
 
-std::string Note::getPitch() const { return getWrittenPitch(); }
+std::string Note::getPitch() const { return getSoundingPitch(); }
 
 bool Note::inChord() const { return _inChord; }