+++ maialib/__init__.py
@@ -1,3 +1,4 @@
 from .maiacore import *
 from .maiapy.other import *
 from .maiapy.plots import *
+from .maiapy.sethares_dissonance import *