+++ maialib/maiapy/sethares_dissonance.py
@@ -1,0 +1,159 @@
+import numpy as np
+import pandas as pd
+import plotly.express as px
+import plotly.graph_objects as go
+from maialib import maiacore as mc
+
+__all__ = ["plotSetharesDissonanceCurve", "plotScoreSetharesDissonance"]
+
+def _dissmeasure(fvec, amp, model='min') -> float:
+    """
+    Given a list of partials in fvec, with amplitudes in amp, this routine
+    calculates the dissonance by summing the roughness of every sine pair
+    based on a model of Plomp-Levelt's roughness curve.
+
+    The older model (model='product') was based on the product of the two
+    amplitudes, but the newer model (model='min') is based on the minimum
+    of the two amplitudes, since this matches the beat frequency amplitude.
+    """
+    # Sort by frequency
+    sort_idx = np.argsort(fvec)
+    am_sorted = np.asarray(amp)[sort_idx]
+    fr_sorted = np.asarray(fvec)[sort_idx]
+
+    # Used to stretch dissonance curve for different freqs:
+    Dstar = 0.24  # Point of maximum dissonance
+    S1 = 0.0207
+    S2 = 18.96
+
+    C1 = 5
+    C2 = -5
+
+    # Plomp-Levelt roughness curve:
+    A1 = -3.51
+    A2 = -5.75
+
+    # Generate all combinations of frequency components
+    idx = np.transpose(np.triu_indices(len(fr_sorted), 1))
+    fr_pairs = fr_sorted[idx]
+    am_pairs = am_sorted[idx]
+
+    Fmin = fr_pairs[:, 0]
+    S = Dstar / (S1 * Fmin + S2)
+    Fdif = fr_pairs[:, 1] - fr_pairs[:, 0]
+
+    if model == 'min':
+        a = np.amin(am_pairs, axis=1)
+    elif model == 'product':
+        a = np.prod(am_pairs, axis=1)  # Older model
+    else:
+        raise ValueError('model should be "min" or "product"')
+    SFdif = S * Fdif
+    D = np.sum(a * (C1 * np.exp(A1 * SFdif) + C2 * np.exp(A2 * SFdif)))
+
+    return D
+
+def plotSetharesDissonanceCurve(base_freq=500, numPartials=6, r_low=1, r_high=2.3, n_points=3000, amplVec=None) -> tuple[go.Figure, list[float], list[float]]:
+    """
+    Compute the Sethares Dissonance Curve of a given base frequency
+
+    Return
+        A tuple that contains a Plotly figure, and the pair 'ratios' and 'dissonance' lists
+    """
+    freqs = base_freq * np.array(list(range(1, numPartials+1)))
+    amps = 0.88**np.array(list(range(0, numPartials))) if amplVec == None else amplVec
+    
+    if len(freqs) != len(amps):
+        raise ValueError("The size of amplVec must be equal to the 'numPartials' (6 is the default)")
+
+    ratios = np.linspace(r_low, r_high, n_points)
+    dissonances = []
+
+    for r in ratios:
+        extended_freqs = np.concatenate([freqs, r * freqs])
+        extended_amps = np.concatenate([amps, amps])
+        d = _dissmeasure(extended_freqs, extended_amps)
+        dissonances.append(d)
+
+    # Plotting using Plotly
+    fig = go.Figure()
+    fig.add_trace(go.Scatter(x=ratios, y=dissonances, mode='lines', name='Dissonance'))
+        
+    # Adding lines for the notable intervals
+    intervals = [(1, 1), (6, 5), (5, 4), (4, 3), (3, 2), (5, 3), (2, 1)]
+    for n, d in intervals:
+        fig.add_shape(
+            type="line",
+            x0=n/d,
+            y0=min(dissonances),
+            x1=n/d,
+            y1=max(dissonances),
+            line=dict(color="Silver", width=1)
+        )
+
+    fig.update_layout(
+        title="Sethares' Sensory Dissonance Curve",
+        xaxis_title="Frequency Ratio",
+        yaxis_title="Sensory Dissonance",
+        xaxis_type="log",
+        xaxis=dict(tickvals=[n/d for n, d in intervals],
+                   ticktext=['{}/{}'.format(n, d) for n, d in intervals]),
+        yaxis=dict(showticklabels=True)
+    )
+    
+    return fig, ratios, dissonances
+
+def plotScoreSetharesDissonance(score: mc.Score, **kwargs) -> tuple[go.Figure, pd.DataFrame]:
+    """Plot 2D line graph of the Sethares Dissonance over time
+
+    Args:
+       score (maialib.Score):  A maialib Score object loaded with a valid MusicXML file
+
+    Kwargs:
+       measureStart (int): Start measure to plot
+       measureEnd (int): End measure to plot
+
+    Returns:
+       A list: [Plotly Figure, The plot data as a Pandas Dataframe]
+
+    Raises:
+       RuntimeError, KeyError
+
+    Examples of use:
+
+    >>> myScore = ml.Score("/path/to/score.xml")
+    >>> ml.plotScoreSetharesDissonance(myScore)
+    >>> ml.plotScoreSetharesDissonance(myScore, measureStart=10, measureEnd=20)
+    """
+    df = score.getChordsDataFrame(kwargs)
+
+    df["dissonance"] = df.apply(lambda row: row.chord.getSetharesDissonanceValue(), axis=1)
+    df["chordNotes"] = df.apply(lambda row: ', '.join([str(x.getPitch()) for x in row.chord.getNotes()]), axis=1)
+    df["chordSize"] = df.apply(lambda row: row.chord.size(), axis=1)
+
+    fig = px.line(df, x="floatMeasure", y="dissonance",
+              hover_data=["chordNotes", "chordSize"],
+              title='<b>Sethares Sensory Dissonance</b><br><i>Beethoven 5th Symphony</i>')
+
+    fig.update_layout(
+        title_x=0.5,
+        xaxis_title="Measures",
+        yaxis_title="Sensory Dissonance",
+    )
+
+    fig.add_shape(
+            # Rectangle with reference to the plot
+            type="rect",
+            xref="paper",
+            yref="paper",
+            x0=0,
+            y0=0,
+            x1=1.0,
+            y1=1.0,
+            line=dict(
+                color="black",
+                width=1,
+            )
+        )
+    
+    return fig, df